Abstracts out deployment execution to allow for extensibility
