+++ src/OrchardCore/OrchardCore.Deployment.Abstractions/IDeploymentTargetHandler.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using Microsoft.Extensions.FileProviders;
+
+namespace OrchardCore.Deployment
+{
+    public interface IDeploymentTargetHandler
+    {
+        Task ImportFromFileAsync(IFileProvider fileProvider);
+    }
+}
